pyproject.toml
[project]
name = "climate-chatbot"
version = "0.1.0"
description = "AI-Powered Climate Policy Assistant"
authors = [
    {name = "Ankit Kumar", email = "ankit@example.com"}
]
requires-python = ">=3.9"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.1.4",
    "groq>=0.4.2",
    "requests>=2.31.0",
    "numpy>=1.26.3"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["db*", "UI*", "etl*", "backend*", "training_data*"]
